The REALbasic Common KFS BSD library got updated.
We must commit the ID of the new HEAD of the library because Git tracks the ID of the HEAD of all submodules.

If the library changes in a way that would break something here, then we would not care until we explicitly tell Git to update the local copy of the library.  Until then, Git will continue to use the library at the commit described by this repository.